Podcast_App/requirements.txt
python-dotenv
spotipy